+++ changelog.md
@@ -40,6 +40,8 @@ All notable changes to this project will be documented in this file.
   ([#35](https://github.com/kotools/types/issues/35)).
 - Refactor binary operations of `kotools.types.number.NegativeInt`
   ([#36](https://github.com/kotools/types/issues/36)).
+- Refactor binary operations of `kotools.types.number.StrictlyNegativeInt`
+  ([#37](https://github.com/kotools/types/issues/37)).
 
 ## Version 1.2.0 - 2022/07/11